requirements.txt
requests
python-dotenv
streamlit
openai
pypdf2
google-generativeai
typing
firebase_admin
fastapi
uvicorn